chore: #2099 update reducer to use combineReducers, fix types on selector
