+++ packages/admin-portal/src/reducers/apps/index.ts
@@ -1,0 +1,10 @@
+import appDetailReducer, { AppDetailState } from './detail'
+import { combineReducers } from 'redux'
+
+export type AppsState = {
+  detail: AppDetailState
+}
+
+export default combineReducers({
+  detail: appDetailReducer,
+})
